iox-#1391 Rename 'scheduler.hpp' to 'posix_scheduler.hpp' and move to new location
