+++ iceoryx_hoofs/BUILD.bazel
@@ -47,6 +47,7 @@ cc_library(
         "posix/ipc/source/*.cpp",
         "posix/sync/source/*.cpp",
         "posix/time/source/*.cpp",
+        "posix/utility/source/*.cpp",
         "posix/vocabulary/source/*.cpp",
         "primitives/source/*.cpp",
         "reporting/source/log/building_blocks/*.cpp",
@@ -75,6 +76,7 @@ cc_library(
         "posix/ipc/include/",
         "posix/sync/include/",
         "posix/time/include/",
+        "posix/utility/include/",
         "posix/vocabulary/include/",
         "primitives/include/",
         "reporting/include/",
