+++ iceoryx_hoofs/README.md
@@ -76,6 +76,7 @@ The module structure is a logical grouping. It is replicated for `concurrent` an
 |`system_configuration` | i        | Collection of free functions which acquire system information like the page-size.                                                                                                                                                          |
 |`UniqueId`             | i        | Monotonic increasing IDs within a process.                                                                                                                                                                                                 |
 |`into`                 | i        |                                                                                                                                                                                                                                            |
+|`Scheduler`            | i        | Supported schedulers and functions to get their priority range are contained here.                                                                                                                                                    |
 
 ### Primitives (primitives)
 
@@ -126,7 +127,6 @@ The module structure is a logical grouping. It is replicated for `concurrent` an
 |`PeriodicTask`         | i        | Periodically executes a callable specified by the template parameter in a configurable time interval.                                                                                                                                 |
 |`smart_lock`           | i        | Creates arbitrary thread-safe constructs which then can be used like smart pointers. If some STL type should be thread safe use the smart_lock to create the thread safe version in one line. Based on some ideas presented in [Wrapping C++ Member Function Calls](https://stroustrup.com/wrapper.pdf) |
 |`mutex`                | i        | Mutex interface, see [ManPage pthread_mutex_lock](https://man7.org/linux/man-pages/man3/pthread_mutex_lock.3p.html).                                                                                                                  |
-|`Scheduler`            |          | Supported schedulers and functions to get their priority range are contained here.                                                                                                                                                    |
 |`UnnamedSemaphore`     |          | Unamed semaphore interface, see [ManPage sem_overview](https://man7.org/linux/man-pages/man7/sem_overview.7.html)                                                                                                                     |
 |`NamedSemaphore`       |          | Named semaphore interface, see [ManPage sem_overview](https://man7.org/linux/man-pages/man7/sem_overview.7.html)                                                                                                                      |
 |`thread`               |          | Heap-less replacement for `std::thread`.                                                                                                                                                                                              |
